app/pyproject.toml
[project]
name = "renew-ai-week-app"
version = "1.0.0"
description = "ReNew Capital Partners - AI Portfolio Intelligence"
requires-python = ">=3.10"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.30.0",
    "httpx>=0.27.0",
    "databricks-sdk>=0.30.0",
    "pydantic>=2.0.0",
]

[build-system]
requires = ["setuptools"]
build-backend = "setuptools.build_meta"
